@@ -1,8 +1,8 @@
-package vald3nir.programming_challenge.services;
+package vald3nir.programming_challenge.control;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
-import vald3nir.programming_challenge.models.DataAssets;
+import vald3nir.programming_challenge.model.DataAssets;
 
 public interface RetrofitServices {
 
